Update to Parsimmon v0.7.2
On the plus side, this gets rid of parsimmon-wrapper.js, which is no
longer necessary. However, they changed the format for tokens from the
character offset in the file to an object also including the line and
column number. That broke parser error messages and the Travis CI build.

Close #3: "Update to latest Parsimmon"