Add an `Integer` primitive.
It's sometimes convenient to specify something as an integer, rather
than supplying a possible range. This implementation is probably not
robust enough and should be backed by bignum, but I'm treating it as an
experiment.

Generalizing, it would be nice to add a primitive `String` type.

I understand the appeal in having a minimal language with a small
surface area, however builtins are rather convenient.

Test plan:
* `npm test`
* Run this contract:
  https://gist.github.com/joelburget/7b9693bb4abde38a046ad2ab3e758e62